e2e: separate tests into 3 categories
This architecture pattern worked quite well in one of the projects of @BacLuc.
