+++ .github/workflows/reusable-e2e-tests-run.yml
@@ -21,6 +21,7 @@ jobs:
           - chromium
           - firefox
           - webkit
+          - behavior-tests
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6
