+++ e2e/README.md
@@ -3,6 +3,30 @@
 As a pre-requisite for running end-to-end tests, we assume you have the application fully up and running on your system.
 If not, please follow the documentation links in the README.md in the root of this repository.
 
+## Structure
+
+The e2e tests are structured into 3 groups of tests:
+
+### [0-snapshot-tests](tests/0-snapshot-tests)
+
+Tests to make sure that views look the same as before.
+**They must not change state** which allows us to run the in parallel.
+They are also run with multiple browsers and viewports to make sure we didn't change something accidentally.
+
+### [5-cross-browser-tests](tests/5-cross-browser-tests)
+
+Tests to make sure that something works in all browsers.
+That can be a hack that we needed for a certain browser (this doesn't happen that often anymore, but it happens)
+or very important behavior we want to be sure it works on all browsers.
+We try to not modify state in this category as long as possible that we can keep them parallelized,
+but they might need to change state at some point.
+
+### [9-behavior-tests](tests/9-behavior-tests)
+
+Tests that we don't have to run in multiple browsers and viewports because they mostly test the interaction between
+the frontend and other services. They must not rely on a certain database state and cannot be run in parallel as
+they might influence the state of other tests.
+
 ## Option A: Run end-to-end tests in Docker container (headless)
 
 ### Preparation
