+++ e2e/tests/5-cross-browser-tests/clientPrint.spec.js
@@ -1,8 +1,8 @@
 // ClientPrint
 
 import { test, expect } from '@playwright/test'
-import { getPdfProperties } from '../utils/getPdfProperties'
-import { loginAndSetCookie } from '../utils/helpers'
+import { getPdfProperties } from '../../utils/getPdfProperties'
+import { loginAndSetCookie } from '../../utils/helpers'
 
 import { readFileSync } from 'fs'
