+++ e2e/tests/9-behavior-tests/httpCache/camps.spec.js
@@ -1,5 +1,5 @@
 import { test } from '@playwright/test'
-import { loginAndSetCookie, expectCachePass } from '../../utils/helpers'
+import { loginAndSetCookie, expectCachePass } from '../../../utils/helpers'
 
 test("doesn't cache /camps", async ({ page, request }) => {
   const uri = '/api/camps'
