+++ e2e/tests/9-behavior-tests/httpCache/content-types.spec.js
@@ -1,13 +1,13 @@
 import { test, expect } from '@playwright/test'
-import collectionResponse from '../../test-data/httpCache/content_types_collection.json'
-import itemResponse from '../../test-data/httpCache/content_types_entity.json'
-import { bipiUser, castorUser } from '../../utils/constants'
+import collectionResponse from '../../../test-data/httpCache/content_types_collection.json'
+import itemResponse from '../../../test-data/httpCache/content_types_entity.json'
+import { bipiUser, castorUser } from '../../../utils/constants'
 import {
   expectCacheHit,
   expectCacheMiss,
   apiGet,
   getAuthContext,
-} from '../../utils/helpers'
+} from '../../../utils/helpers'
 
 const collectionXKeys =
   'a4211c11211c f17470519474 1a0f84e322c8 c462edd869f3 5e2028c55ee4 3ef17bd1df72 4f0c657fecef a4211c112939 44dcc7493c65 cfccaecd4bad 318e064ea0c9 /api/content_types'
