+++ e2e/tests/9-behavior-tests/httpCache/root.spec.js
@@ -1,4 +1,8 @@
-import { expectCacheHit, expectCacheMiss, loginAndSetCookie } from '../../utils/helpers'
+import {
+  expectCacheHit,
+  expectCacheMiss,
+  loginAndSetCookie,
+} from '../../../utils/helpers'
 import { test } from '@playwright/test'
 
 const user1 = 'test@example.com'
